refactor: Move database type management to db module.
