+++ test/orbitdb-open.test.js
@@ -4,7 +4,7 @@ import fs from 'fs'
 import path from 'path'
 import * as IPFS from 'ipfs-core'
 import { OrbitDB, isValidAddress, LevelStorage } from '../src/index.js'
-import { KeyValueIndexed } from '../src/db/index.js'
+import KeyValueIndexed from '../src/db/keyvalue-indexed.js'
 import config from './config.js'
 import connectPeers from './utils/connect-nodes.js'
 import waitFor from './utils/wait-for.js'
